+++ src/routes/utasi.route.ts
@@ -1,33 +1,36 @@
-import { TrainAssetsController } from "../controllers/train.assets.controller";
+import { TrainController } from "../controllers/train.controller";
 import { UTASIController } from "../controllers/utasi.controller";
+import { ExternalController } from "../controllers/external.controller";
+import { StationController } from "../controllers/station.controller";
 import express from "express";
+import { AvailabilityController } from "../controllers/availability.controller";
 
 const router = express.Router();
 
+// UTASI Routes
 router.get("/contracts", UTASIController.getContracts);
-
 router.post("/contracts/attach", UTASIController.attachContract);
-
-router.put("/asset/:id", UTASIController.updateAsset);
-
-router.put("/parapets/status/update", UTASIController.updateParapetStatus);
-
-router.get("/stations", UTASIController.getStations);
-
-router.get("/stations/details", UTASIController.getAllStationDetails);
-
-router.get("/stations/specs", UTASIController.getStationSpecs);
-
 router.get("/stations/contracts", UTASIController.getContractFromAsset);
+router.delete("/contracts/:id", UTASIController.untagContract);
 
-router.get("/train/assets", TrainAssetsController.getTrainAssets);
+// Station Routes
+router.put("/asset/:id", StationController.updateAsset);
+router.put("/parapets/status/update", StationController.updateParapetStatus);
+router.get("/stations", StationController.getStations);
+router.get("/stations/details", StationController.getAllStationDetails);
+router.get("/stations/specs", StationController.getStationSpecs);
 
-router.get("/train/assetsSpecs", TrainAssetsController.getTrainAssetSpecs);
+// Availability
+router.get("/availability/parapets", AvailabilityController.getParapetsAvailability);
+router.get("/availability/backlits", AvailabilityController.getBacklitsAvailability );
 
-router.get("/train/external/specs", TrainAssetsController.getExternalAssetSpecs);
+// Train Routes
+router.get("/train/assets", TrainController.getTrainAssets);
+router.get("/train/assetsSpecs", TrainController.getTrainAssetSpecs);
+router.put("/train/assets/book/:id", TrainController.bookTrainAsset);
+router.put("/train/assets/edit/:id", TrainController.updateTrainAsset);
 
-router.put("/train/assets/book/:id", TrainAssetsController.bookTrainAsset);
-
-router.put("/train/assets/edit/:id", TrainAssetsController.updateTrainAsset);
+// External Routes
+router.get("/train/external/specs/:asset_id", ExternalController.getExternalAssetSpecs);
 
 export const UTASIRoute = router;
